+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Adminer dev
+- Autocomplete SQL commands
 - Do not edit NULL values by Modify (bug #967)
 - PostgreSQL: Support COPY FROM stdin in SQL query (bug #942)
 - MySQL: Display number of found rows in group queries (regression from 5.1.1)